fix docstring test import after refactor
